Rename new tab files to match formatting guidelines
Rename new tab files to match formatting guidelines
