+++ lib/ui/home/HomePage.dart
@@ -7,8 +7,8 @@ import 'package:cobble/ui/common/icons/WatchIcon.dart';
 import 'package:cobble/ui/common/icons/fonts/RebbleIconsStroke.dart';
 import 'package:cobble/ui/home/tabs/StoreTab.dart';
 import 'package:cobble/ui/home/tabs/TestTab.dart';
-import 'package:cobble/ui/home/tabs/SettingsTab.dart';
-import 'package:cobble/ui/home/tabs/MyWatchesTab.dart';
+import 'package:cobble/ui/home/tabs/settings_tab.dart';
+import 'package:cobble/ui/home/tabs/watches_tab.dart';
 import 'package:cobble/ui/setup/FirstRunPage.dart';
 import 'package:cobble/ui/setup/PairPage.dart';
 import 'package:cobble/ui/test/WatchCarousel.dart';
